move trace lifecycle into run pkg
